Refactor out local auth functions
